Starting to rewrite for scalability.
Restructured almost everything.
Browser clients need to select the robot they are interested in.
Browser currently gets notifications when any robot connects or disconnects (that might change if it gets to a point where there are enough simultaneous clients for it to become a problem).
Need to create a view for requesting current list of robots available for pre-populating the robots list.